move AuthModal to AuthContext
